+++ P_dev_model/P_FRQmetter_signal/processing_FRQmetter_extern.h
@@ -16,9 +16,7 @@
 #define DEV_5                FRQmetter
 //#define DEV_1_BIT_STATUS_REG 1
 
-// Количекство статус-регистров FRQmetter
 #define NUM_REG_STATUS_TC       1
-// Общее количество регистров FRQmetter (имееться в виду регистров разных типов, без status регистра)
 #define NUM_REG_TC              1
 
 typedef enum{
@@ -31,28 +29,21 @@ typedef enum{
 
 #pragma pack(push,1)
 
-//-----------------------------Адреса оперативных регистров процесса FRQmetter---------------------------------------------------------------
 typedef struct{
-	uint16_t  status_FRQmetter;  // статус регистры FRQmetter
-	int16_t   rez_FRQmetter;     // адрес в памяти МК регистров состояния FRQmetter
+	uint16_t  status_FRQmetter;
+	int16_t   rez_FRQmetter;
 } S_FRQmetter_address;
-//--------------------------------------------------------------------------------------------------------------------------------------
 
-
-//----------------------------- Оперативные регистры процесса FRQmetter------------------------------------------------------------------------
 typedef struct{
-	S_proces_object_modbus  status_FRQmetter; // статус регистры FRQmetter
-	S_proces_object_modbus  rez_FRQmetter;    // measurement result
+	S_proces_object_modbus  status_FRQmetter;
+	S_proces_object_modbus  rez_FRQmetter;
 } S_FRQmetter_oper_data;
-//--------------------------------------------------------------------------------------------------------------------------------------
 
-
-//----------------------------- Конфигурация процесса FRQmetter--------------------------------------------------------------------------------
 typedef struct{
-	uint16_t state;           // state of module ENABLE/DISABLE
+	uint8_t state;           // state of module ENABLE/DISABLE
 	int16_t  frqCorrection;   // correction of measured frequency: 1 = 1/100 Hz
 }S_FRQmetter_user_config;
-//--------------------------------------------------------------------------------------------------------------------------------------
+
 #pragma pack(pop)
 
 uint16_t FRQmetter_calc_address_oper_reg(S_FRQmetter_address *ps_FRQmetter_address, uint16_t adres_start);
